+++ projects/nb-trans/src/lib/models/nb-trans-params.interface.ts
@@ -1,3 +1,8 @@
+/**
+ * The naming rules about param key:
+ * 1. Consists of letters, numbers, _, and $
+ * 2. The number can't be the first character
+ */
 export interface INbTransParams {
   [key: string]: string;
 }